Refactored ads.aqua.model module
