+++ ads/aqua/deployment.py
@@ -32,9 +32,10 @@ from ads.aqua.data import (
     InferenceContainerParamType,
     InferenceContainerType,
     InferenceContainerTypeKey,
+    Tags,
 )
 from ads.aqua.finetune import FineTuneCustomMetadata
-from ads.aqua.model import AquaModelApp, Tags
+from ads.aqua.model import AquaModelApp
 from ads.common.auth import default_signer
 from ads.common.object_storage_details import ObjectStorageDetails
 from ads.common.serializer import DataClassSerializable
