+++ ads/aqua/extension/model_handler.py
@@ -19,8 +19,9 @@ from tornado.web import HTTPError
 from ads.aqua.common.decorator import handle_exceptions
 from ads.aqua.common.errors import AquaRuntimeError
 from ads.aqua.extension.base_handler import AquaAPIhandler, Errors
-from ads.aqua.extension.utils import validate_function_parameters
-from ads.aqua.model import AquaModelApp, AquaModelSummary, HFModelSummary, ModelTask
+from ads.aqua.model import AquaModelApp
+from ads.aqua.model.constants import ModelTask
+from ads.aqua.model.entities import AquaModelSummary, HFModelSummary
 
 
 class AquaModelHandler(AquaAPIhandler):
